backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.1

# CORS
python-multipart==0.0.20

# Data Processing (for seeding from parquet)
pandas==2.2.3
pyarrow==18.1.0

# Authentication
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing
